requirements.txt
python-binance
numpy
TA-Lib
unicorn-binance-websocket-api
